# src/models/ProviderCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Torrents Stream Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../helpers';
import {
    ProviderCategorySubcategories,
    ProviderCategorySubcategoriesFromJSON,
    ProviderCategorySubcategoriesFromJSONTyped,
    ProviderCategorySubcategoriesToJSON,
} from './';

/**
 * 
 * @export
 * @interface ProviderCategory
 */
export interface ProviderCategory {
    /**
     * 
     * @type {string}
     * @memberof ProviderCategory
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ProviderCategory
     */
    id: string;
    /**
     * 
     * @type {Array<ProviderCategorySubcategories>}
     * @memberof ProviderCategory
     */
    subcategories: Array<ProviderCategorySubcategories>;
}

export function ProviderCategoryFromJSON(json: any): ProviderCategory {
    return ProviderCategoryFromJSONTyped(json, false);
}

export function ProviderCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProviderCategory {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'id': json['id'],
        'subcategories': ((json['subcategories'] as Array<any>).map(ProviderCategorySubcategoriesFromJSON)),
    };
}

export function ProviderCategoryToJSON(value?: ProviderCategory | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'id': value.id,
        'subcategories': ((value.subcategories as Array<any>).map(ProviderCategorySubcategoriesToJSON)),
    };
}