# src/models/ProviderTorrent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Torrents Stream Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../helpers';
/**
 * 
 * @export
 * @interface ProviderTorrent
 */
export interface ProviderTorrent {
    /**
     * 
     * @type {string}
     * @memberof ProviderTorrent
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ProviderTorrent
     */
    magnet: string;
    /**
     * 
     * @type {number}
     * @memberof ProviderTorrent
     */
    seeds: number;
    /**
     * 
     * @type {number}
     * @memberof ProviderTorrent
     */
    downloads?: number;
    /**
     * 
     * @type {number}
     * @memberof ProviderTorrent
     */
    peers: number;
    /**
     * 
     * @type {string}
     * @memberof ProviderTorrent
     */
    size: string;
    /**
     * 
     * @type {string}
     * @memberof ProviderTorrent
     */
    time?: string;
}

export function ProviderTorrentFromJSON(json: any): ProviderTorrent {
    return ProviderTorrentFromJSONTyped(json, false);
}

export function ProviderTorrentFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProviderTorrent {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'magnet': json['magnet'],
        'seeds': json['seeds'],
        'downloads': !exists(json, 'downloads') ? undefined : json['downloads'],
        'peers': json['peers'],
        'size': json['size'],
        'time': !exists(json, 'time') ? undefined : json['time'],
    };
}

export function ProviderTorrentToJSON(value?: ProviderTorrent | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'magnet': value.magnet,
        'seeds': value.seeds,
        'downloads': value.downloads,
        'peers': value.peers,
        'size': value.size,
        'time': value.time,
    };
}